+++ src/core/Reuni.ts
@@ -6,6 +6,7 @@ import {
   KeyCareItem,
   StoreValidDict,
   Observer,
+  StoreObserver,
   DirtyNodes
 } from "./types";
 import { genId } from "./utils";
@@ -15,13 +16,14 @@ import TaskManager from "./TaskManager";
 import Node from "./Node";
 import TaskHandler from "../api/TaskHandler";
 import { ObserverCB } from "../api/types";
-import ObManager from "./ObManager";
+import NodeObManager from "./NodeObManager";
+import StoreObManager from "./StoreObManager";
 
 export default class Reuni {
   private _nodeDict: NodeDict;
   private _rootId: string;
-  private _storeObs: ObManager;
-  private _observers: ObManager;
+  private _storeObs: StoreObManager;
+  private _observers: NodeObManager;
   private _storeValidDict: StoreValidDict;
   private _dirtyNodes: DirtyNodes;
   private _taskManager: TaskManager;
@@ -41,8 +43,8 @@ export default class Reuni {
       }
     };
     this._taskManager = new TaskManager();
-    this._observers = new ObManager();
-    this._storeObs = new ObManager();
+    this._observers = new NodeObManager();
+    this._storeObs = new StoreObManager();
     this._dirtyNodes = {};
     this._storeValidDict = {};
   }
@@ -132,12 +134,7 @@ export default class Reuni {
     let store = node.ref.addStore(storeName, RawStore, storeCare);
     storeValidDict[storeName] = store;
     let beValidObs = this._storeObs.addStoreRefresh(nodeId, storeName, this);
-    beValidObs.every(ob => {
-      if (ob.storeName != null) {
-        this._observers.enableStoreNotify(ob.nodeId, ob.storeName, this) != null;
-      }
-      return this._observers.getInvalidObs().length !== 0;
-    });
+    this._observers.enableStores(beValidObs, this);
     return store;
   }
 
@@ -159,13 +156,7 @@ export default class Reuni {
     let beInvalidObs = this._storeObs
       .deleteStoreRefresh(nodeId, storeName, this)
       .concat(storeOb);
-    beInvalidObs.every(ob => {
-      if (ob.storeName != null) {
-        this._observers.disableStoreNotify(ob.nodeId, ob.storeName, this) !=
-          null;
-      }
-      return this._observers.getObs().length !== 0;
-    });
+    this._observers.disableStores(beInvalidObs, this);
     return store;
   }
 
@@ -192,7 +183,7 @@ export default class Reuni {
   }
 
   observe(care: ObserverCareDict, nodeId: string, cb: ObserverCB) {
-    return this._observers.observe(care, nodeId, null, cb, this);
+    return this._observers.observe(care, nodeId, cb, this);
   }
 
   storeObserve(
@@ -204,7 +195,7 @@ export default class Reuni {
     return this._storeObs.observe(care, nodeId, storeName, cb, this);
   }
 
-  storeUnobserve(observer: Observer) {
+  storeUnobserve(observer: StoreObserver) {
     this._storeObs.unobserve(observer);
   }
